updated model to use doses and decimals

+++ lib/data/providers/medication_intake_provider.dart
@@ -1,3 +1,4 @@
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import '../model/medication_intake.dart';
 import '../repositories/medication_intake_repository.dart';
@@ -38,7 +39,7 @@ class MedicationIntakeProvider extends ChangeNotifier {
     fetchIntakes();
   }
 
-  Future<void> addIntake(DateTime scheduledDateTime, double dose) async {
+  Future<void> addIntake(DateTime scheduledDateTime, Decimal dose) async {
     await MedicationIntakeRepository.insertIntake(MedicationIntake(
         scheduledDateTime: scheduledDateTime, dose: dose));
     fetchIntakes();
